improve task messages render
