+++ src/features/Conversation/Messages/Task/ClientTaskDetail/index.tsx
@@ -1,7 +1,6 @@
 'use client';
 
 import { type TaskDetail, ThreadStatus } from '@lobechat/types';
-import { Flexbox } from '@lobehub/ui';
 import { memo, useMemo } from 'react';
 
 import BubblesLoading from '@/components/BubblesLoading';
@@ -9,10 +8,8 @@ import { useChatStore } from '@/store/chat';
 import { displayMessageSelectors } from '@/store/chat/selectors';
 import { messageMapKey } from '@/store/chat/utils/messageMapKey';
 
-import CompletedState from './CompletedState';
+import { TaskMessages } from '../../Tasks/shared';
 import InitializingState from './InitializingState';
-import InstructionAccordion from './InstructionAccordion';
-import ProcessingState from './ProcessingState';
 
 interface ClientTaskDetailProps {
   /** Agent ID from the task message (use task's agentId, not activeAgentId) */
@@ -24,92 +21,76 @@ interface ClientTaskDetailProps {
   taskDetail?: TaskDetail;
 }
 
-const ClientTaskDetail = memo<ClientTaskDetailProps>(({ agentId: propAgentId, groupId, taskDetail }) => {
-  const threadId = taskDetail?.threadId;
-  const isExecuting = taskDetail?.status === ThreadStatus.Processing;
+const ClientTaskDetail = memo<ClientTaskDetailProps>(
+  ({ agentId: propAgentId, groupId, taskDetail }) => {
+    const threadId = taskDetail?.threadId;
+    const isExecuting = taskDetail?.status === ThreadStatus.Processing;
 
-  // Use task message's agentId to query with the correct SubAgent ID that created the thread
-  // Fall back to activeAgentId if task message doesn't have agentId (shouldn't happen normally)
-  const [activeAgentId, activeTopicId, useFetchMessages] = useChatStore((s) => [
-    s.activeAgentId,
-    s.activeTopicId,
-    s.useFetchMessages,
-  ]);
+    // Use task message's agentId to query with the correct SubAgent ID that created the thread
+    // Fall back to activeAgentId if task message doesn't have agentId (shouldn't happen normally)
+    const [activeAgentId, activeTopicId, useFetchMessages] = useChatStore((s) => [
+      s.activeAgentId,
+      s.activeTopicId,
+      s.useFetchMessages,
+    ]);
 
-  const agentId = propAgentId || activeAgentId;
+    const agentId = propAgentId || activeAgentId;
 
-  const threadContext = useMemo(
-    () => ({
-      agentId,
-      groupId,
-      scope: 'thread' as const,
-      threadId,
-      topicId: activeTopicId,
-    }),
-    [agentId, groupId, activeTopicId, threadId],
-  );
+    const threadContext = useMemo(
+      () => ({
+        agentId,
+        groupId,
+        scope: 'thread' as const,
+        threadId,
+        topicId: activeTopicId,
+      }),
+      [agentId, groupId, activeTopicId, threadId],
+    );
 
-  const threadMessageKey = useMemo(
-    () => (threadId ? messageMapKey(threadContext) : null),
-    [threadId],
-  );
+    const threadMessageKey = useMemo(
+      () => (threadId ? messageMapKey(threadContext) : null),
+      [threadId],
+    );
 
-  // Fetch thread messages (skip when executing - messages come from real-time updates)
-  useFetchMessages(threadContext, isExecuting);
+    // Fetch thread messages (skip when executing - messages come from real-time updates)
+    useFetchMessages(threadContext, isExecuting);
 
-  // Get thread messages from store using selector
-  const threadMessages = useChatStore((s) =>
-    threadMessageKey
-      ? displayMessageSelectors.getDisplayMessagesByKey(threadMessageKey)(s)
-      : undefined,
-  );
+    // Get thread messages from store using selector
+    const threadMessages = useChatStore((s) =>
+      threadMessageKey
+        ? displayMessageSelectors.getDisplayMessagesByKey(threadMessageKey)(s)
+        : undefined,
+    );
 
-  if (!threadMessages) return <BubblesLoading />;
+    if (!threadMessages) return <BubblesLoading />;
 
-  // Find the assistantGroup message which contains the children blocks
-  const assistantGroupMessage = threadMessages.find((item) => item.role === 'assistantGroup');
-  const instruction = threadMessages.find((item) => item.role === 'user')?.content;
-  const childrenCount = assistantGroupMessage?.children?.length ?? 0;
+    // Find the assistantGroup message which contains the children blocks
+    const assistantGroupMessage = threadMessages.find((item) => item.role === 'assistantGroup');
+    const childrenCount = assistantGroupMessage?.children?.length ?? 0;
 
-  // Get model/provider from assistantGroup message
-  const model = assistantGroupMessage?.model;
-  const provider = assistantGroupMessage?.provider;
+    // Get model/provider from assistantGroup message
+    const model = assistantGroupMessage?.model;
+    const provider = assistantGroupMessage?.provider;
 
-  // Initializing state: no status yet (task just created, waiting for client execution)
-  if (threadMessages.length === 0 || !assistantGroupMessage?.children || childrenCount === 0) {
-    return <InitializingState />;
-  }
+    // Initializing state: no status yet (task just created, waiting for client execution)
+    if (threadMessages.length === 0 || !assistantGroupMessage?.children || childrenCount === 0) {
+      return <InitializingState />;
+    }
 
-  return (
-    <Flexbox gap={4}>
-      {instruction && (
-        <InstructionAccordion childrenCount={childrenCount} instruction={instruction} />
-      )}
+    return (
+      <TaskMessages
+        duration={taskDetail?.duration}
+        isProcessing={isExecuting}
+        messages={threadMessages}
+        model={model ?? undefined}
+        provider={provider ?? undefined}
+        startTime={assistantGroupMessage.createdAt}
+        totalCost={taskDetail?.totalCost}
+      />
+    );
+  },
+);
 
-      {isExecuting ? (
-        <ProcessingState
-          assistantId={assistantGroupMessage.id}
-          blocks={assistantGroupMessage.children}
-          model={model ?? undefined}
-          provider={provider ?? undefined}
-          startTime={assistantGroupMessage.createdAt}
-        />
-      ) : (
-        <CompletedState
-          assistantId={assistantGroupMessage.id}
-          blocks={assistantGroupMessage.children}
-          duration={taskDetail?.duration}
-          model={model ?? undefined}
-          provider={provider ?? undefined}
-          totalCost={taskDetail?.totalCost}
-          totalTokens={taskDetail?.totalTokens}
-          totalToolCalls={taskDetail?.totalToolCalls}
-        />
-      )}
-    </Flexbox>
-  );
-});
-
-ClientTaskDetail.displayName = 'ClientClientTaskDetail';
+ClientTaskDetail.displayName = 'ClientTaskDetail';
 
 export default ClientTaskDetail;
