+++ src/features/Conversation/Messages/Task/TaskDetailPanel/StatusContent.tsx
@@ -2,13 +2,13 @@
 
 import { memo } from 'react';
 
+import { useChatStore } from '@/store/chat';
 import { type TaskDetail, ThreadStatus } from '@/types/index';
 
 import {
-  CompletedState,
   ErrorState,
   InitializingState,
-  ProcessingState,
+  TaskMessages,
   isProcessingStatus,
 } from '../../Tasks/shared';
 
@@ -18,22 +18,35 @@ interface StatusContentProps {
   taskDetail?: TaskDetail;
 }
 
-const StatusContent = memo<StatusContentProps>(({ taskDetail, content, messageId }) => {
+const StatusContent = memo<StatusContentProps>(({ taskDetail, messageId }) => {
   const status = taskDetail?.status;
+  const threadId = taskDetail?.threadId;
+  const isProcessing = isProcessingStatus(status);
+
+  // Get polling hook - poll for task status to get messages
+  const [useEnablePollingTaskStatus, operations] = useChatStore((s) => [
+    s.useEnablePollingTaskStatus,
+    s.operations,
+  ]);
+
+  // Check if exec_async_task is already polling for this message
+  const hasActiveOperationPolling = Object.values(operations).some(
+    (op) =>
+      op.status === 'running' &&
+      op.type === 'execAgentRuntime' &&
+      op.context?.messageId === messageId,
+  );
+
+  // Enable polling when task has threadId and no active operation is polling
+  // For completed tasks, this will fetch messages once (no refreshInterval needed)
+  const shouldPoll = !!threadId && !hasActiveOperationPolling;
+  const { data } = useEnablePollingTaskStatus(threadId, messageId, shouldPoll);
+
+  const messages = data?.messages;
 
   // Initializing state: no status yet (task just created, waiting for backend)
   if (!status) {
     return <InitializingState />;
-  }
-
-  // Processing states: Processing, InReview, Pending, Active, Todo
-  if (isProcessingStatus(status)) {
-    return <ProcessingState messageId={messageId} taskDetail={taskDetail!} variant="detail" />;
-  }
-
-  // Completed state
-  if (status === ThreadStatus.Completed) {
-    return <CompletedState content={content} taskDetail={taskDetail!} variant="detail" />;
   }
 
   // Error states: Failed, Cancel
@@ -41,7 +54,20 @@ const StatusContent = memo<StatusContentProps>(({ taskDetail, content, messageId
     return <ErrorState taskDetail={taskDetail!} />;
   }
 
-  // Fallback to initializing state for unknown status
+  // Processing or Completed state with messages
+  if (messages && messages.length > 0) {
+    return (
+      <TaskMessages
+        duration={taskDetail?.duration}
+        isProcessing={isProcessing}
+        messages={messages}
+        startTime={taskDetail?.startedAt ? new Date(taskDetail.startedAt).getTime() : undefined}
+        totalCost={taskDetail?.totalCost}
+      />
+    );
+  }
+
+  // Still loading messages
   return <InitializingState />;
 });
