+++ src/features/Conversation/Messages/Tasks/TaskItem/ClientTaskItem.tsx
@@ -9,9 +9,7 @@ import { messageMapKey } from '@/store/chat/utils/messageMapKey';
 import { ThreadStatus } from '@/types/index';
 import type { UIChatMessage } from '@/types/index';
 
-import ClientTaskDetailCompletedState from '../../Task/ClientTaskDetail/CompletedState';
-import ClientTaskDetailProcessingState from '../../Task/ClientTaskDetail/ProcessingState';
-import { ErrorState, InitializingState, isProcessingStatus } from '../shared';
+import { ErrorState, InitializingState, TaskMessages, isProcessingStatus } from '../shared';
 import TaskTitle, { type TaskMetrics } from './TaskTitle';
 
 interface ClientTaskItemProps {
@@ -43,8 +41,7 @@ const ClientTaskItem = memo<ClientTaskItemProps>(({ item }) => {
 
   // Use task message's agentId (skip 'supervisor' as it's not a valid agent ID for queries)
   // Fall back to activeAgentId if not available
-  const agentId =
-    itemAgentId && itemAgentId !== 'supervisor' ? itemAgentId : activeAgentId;
+  const agentId = itemAgentId && itemAgentId !== 'supervisor' ? itemAgentId : activeAgentId;
 
   const threadContext = useMemo(
     () => ({
@@ -135,33 +132,21 @@ const ClientTaskItem = memo<ClientTaskItemProps>(({ item }) => {
         {/* Initializing State - no taskDetail yet or no blocks */}
         {(isInitializing || (isProcessing && !hasBlocks)) && <InitializingState />}
 
-        {/* Processing State - show streaming blocks */}
-        {!isInitializing && isProcessing && hasBlocks && (
-          <ClientTaskDetailProcessingState
-            assistantId={assistantGroupMessage!.id}
-            blocks={blocks!}
+        {/* Processing or Completed State - show blocks via TaskMessages */}
+        {!isInitializing && (isProcessing || isCompleted) && hasBlocks && threadMessages && (
+          <TaskMessages
+            duration={taskDetail?.duration}
+            isProcessing={isProcessing}
+            messages={threadMessages}
             model={model ?? undefined}
             provider={provider ?? undefined}
             startTime={assistantGroupMessage?.createdAt}
+            totalCost={taskDetail?.totalCost}
           />
         )}
 
         {/* Error State */}
         {!isInitializing && isError && taskDetail && <ErrorState taskDetail={taskDetail} />}
-
-        {/* Completed State - show blocks with final result */}
-        {!isInitializing && isCompleted && taskDetail && hasBlocks && (
-          <ClientTaskDetailCompletedState
-            assistantId={assistantGroupMessage!.id}
-            blocks={blocks!}
-            duration={taskDetail.duration}
-            model={model ?? undefined}
-            provider={provider ?? undefined}
-            totalCost={taskDetail.totalCost}
-            totalTokens={taskDetail.totalTokens}
-            totalToolCalls={taskDetail.totalToolCalls}
-          />
-        )}
       </Block>
     </AccordionItem>
   );
